disable logging during tests, split tests.py file
