+++ code/api/tests/test_authors.py
@@ -5,6 +5,7 @@ from django.test import TestCase
 from django.urls import reverse
 from mixer.backend.django import mixer
 import json
+import logging
 
 from socialDistribution.models import LocalAuthor
 
@@ -19,6 +20,19 @@ def create_author(id, username, displayName, githubUrl, profileImageUrl):
 
 
 class AuthorsViewTests(TestCase):
+    
+    # the pillow, https://stackoverflow.com/users/2812257/the-pillow, "How can I disable logging while running unit tests in Python Django?"
+    # https://stackoverflow.com/a/54519433, 2019-02-04, CC BY-SA 4.0
+
+    # disable logging before tests
+    @classmethod
+    def setUpClass(cls):
+        logging.disable(logging.CRITICAL)
+
+    # enable logging after tests
+    @classmethod
+    def tearDownClass(cls):
+        logging.disable(logging.NOTSET)
 
     def test_get_authors_basic(self):
         expected = json.dumps({"type": "authors", "items": []})
