+++ code/api/tests/test_comments.py
@@ -4,10 +4,11 @@ from django.contrib.auth.models import User
 from django.test import TestCase
 from django.urls import reverse
 from django.utils import timezone
-import datetime
-
 from mixer.backend.django import mixer
+
 import json
+import datetime
+import logging
 
 from socialDistribution.models import Comment
 from .test_authors import create_author
@@ -32,6 +33,19 @@ def get_comments_json(post, comments, page=None, size=None):
         "comments": comments
     }
 class PostCommentsViewTest(TestCase):
+
+    # the pillow, https://stackoverflow.com/users/2812257/the-pillow, "How can I disable logging while running unit tests in Python Django?"
+    # https://stackoverflow.com/a/54519433, 2019-02-04, CC BY-SA 4.0
+
+    # disable logging before tests
+    @classmethod
+    def setUpClass(cls):
+        logging.disable(logging.CRITICAL)
+
+    # enable logging after tests
+    @classmethod
+    def tearDownClass(cls):
+        logging.disable(logging.NOTSET)
 
     def test_get_comment_basic(self):
         post = mixer.blend('socialDistribution.localpost')
