+++ code/api/tests/test_inbox.py
@@ -4,12 +4,14 @@ from django.contrib.auth.models import User
 from django.test import TestCase
 from django.urls import reverse
 from mixer.backend.django import mixer
+
 from datetime import datetime, timezone
+import base64
+import logging
 
 from socialDistribution.models import LocalAuthor, LocalPost, Comment
 from api.models import Node
 from cmput404.constants import *
-import base64
 
 # Documentation and code samples taken from the following references:
 # Django Software Foundation, https://docs.djangoproject.com/en/3.2/intro/tutorial05/
@@ -25,6 +27,19 @@ def create_author(id, username, displayName, githubUrl):
 
 
 class InboxViewTests(TestCase):
+
+    # the pillow, https://stackoverflow.com/users/2812257/the-pillow, "How can I disable logging while running unit tests in Python Django?"
+    # https://stackoverflow.com/a/54519433, 2019-02-04, CC BY-SA 4.0
+
+    # disable logging before tests
+    @classmethod
+    def setUpClass(cls):
+        logging.disable(logging.CRITICAL)
+
+    # enable logging after tests
+    @classmethod
+    def tearDownClass(cls):
+        logging.disable(logging.NOTSET)
 
     def setUp(self):
         Node.objects.create(host=HOST, username='testclient', password='testpassword!', remote_credentials=False)
