+++ code/api/tests/test_posts.py
@@ -5,12 +5,14 @@ from django.test import TestCase
 from django.urls import reverse
 from django.utils import timezone
 from mixer.backend.django import mixer
+
 import base64 as b64
+import logging
+from datetime import datetime
 
 from socialDistribution.models import LocalPost, Category, LocalAuthor
 from .test_authors import create_author
 from cmput404.constants import HOST, API_PREFIX
-from datetime import datetime
 
 def create_post(title, author):
     post = LocalPost.objects.create(
@@ -52,6 +54,19 @@ def get_post_json(post):
         }
 
 class PostsViewTest(TestCase):
+
+    # the pillow, https://stackoverflow.com/users/2812257/the-pillow, "How can I disable logging while running unit tests in Python Django?"
+    # https://stackoverflow.com/a/54519433, 2019-02-04, CC BY-SA 4.0
+
+    # disable logging before tests
+    @classmethod
+    def setUpClass(cls):
+        logging.disable(logging.CRITICAL)
+
+    # enable logging after tests
+    @classmethod
+    def tearDownClass(cls):
+        logging.disable(logging.NOTSET)
 
     def test_get_posts_basic(self):
         self.maxDiff = None
