+++ code/socialDistribution/tests/__init__.py
@@ -1,0 +1,2 @@
+# David Lam, https://stackoverflow.com/users/1093174/david-lam, FlyingZipper, https://stackoverflow.com/users/7506938/flyingzipper, 
+# "Django test runner not finding tests", 2021-01-07, https://stackoverflow.com/a/25820203, CC BY-SA 4.0
